Fix copy/paste feature
