+++ src/utils/index.js
@@ -55,10 +55,23 @@ function generateNodeName(type, paletteLabel, existingNodes) {
   );
   return `${tmpName}_${nodesWithSameName.length + 1}`;
 }
+function processNewNode( node, nodes ) {
+	const { id: nodeId, description, label, parameters } = node;
+	const { type } = parameters;
+	const generatedLabel = generateNodeName(type, label, nodes);
+	return {
+		label: generatedLabel,
+		nodeId,
+		type,
+		description,
+		isValid: false,
+	};
+}
 
 export {
   isNodeLabelValid,
   getImmediateUpstreamNodes,
   getImmediateDownstreamNodes,
   generateNodeName,
+  processNewNode
 };
